reckon is no longer required on units

+++ core/text/src/lib.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 pub mod html;
 pub mod plain;
 
-pub const PLAIN: u64 = 1;
+pub const PLAIN: u16 = 1;
 
 #[macro_use]
 extern crate macro_rules_attribute;
